requirements.txt
streamlit
openpyxl
playwright
pillow
pandas